% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false.R
\name{false}
\alias{false}
\title{A filter object representing a contradiction}
\usage{
false()
}
\value{
A filter object which is \code{FALSE} for all rows
}
\description{
\code{true} returns a filter object representing a contradiction.
}
