% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterr.R
\docType{package}
\name{filterr}
\alias{filterr}
\alias{filterr-package}
\title{filterr: A package for filtering data programmatically}
\description{
The filterr package provides objects representing logical
conditions on rows of data.
The objects allow a user to apply the same filtering query
to different data sets.
}
