% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{is_column}
\alias{is_column}
\title{A function indicating if an object is a column}
\usage{
is_column(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE} / \code{FALSE} iff \code{x} is a column
}
\description{
\code{is_column} indicates if an object is a column.
}
