% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{is_filter_object}
\alias{is_filter_object}
\title{Indicator function for filter objects}
\usage{
is_filter_object(x)
}
\arguments{
\item{x}{Object}
}
\value{
\code{TRUE}/\code{FALSE}
}
\description{
\code{is_filter_object} indicates if a given object is a
filter object.
}
