% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.R
\docType{methods}
\name{not_filter}
\alias{not_filter}
\alias{!,TrueFilterObject-method}
\alias{!,FalseFilterObject-method}
\title{Negation operation}
\usage{
\S4method{!}{TrueFilterObject}(x)

\S4method{!}{FalseFilterObject}(x)
}
\arguments{
\item{x}{Filter object}
}
\description{
Negation operation
}
