% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true.R
\name{true}
\alias{true}
\title{A filter object representing a tautology}
\usage{
true()
}
\value{
A filter object which is \code{TRUE} for all rows
}
\description{
\code{true} returns a filter object representing a tautology.
}
